fix oom issue when qgc is in background
The oom is caused by decoder output buffer not released in time.
Root cause is output buffers are post to event loop of qgc but QT's
implementation stops handling any event if app is in background.

So we just stop to post buffer for rendering in the case to fix it.

Change-Id: Iab93f8dd5d32ef36d89f103feaae144acaaf5abb
Signed-off-by: mazhiqiang <mazhiqiang@pinecone.net>